Extract type def generation to separate file
